+++ CHANGES.md
@@ -10,6 +10,7 @@ CHANGELOG
 * Radial menu to interact with the objects
 * Optionnaly add an empty area surrounding a point
 * Move a point
+* Rotate a point
 * Improve the line/polygon vertex edition
 
 0.5